+++ pkg/gm_geometry/gm_point.go
@@ -1,6 +1,6 @@
 package gm_geometry
 
 type Point struct {
-	X float32
-	Y float32
+	X float64
+	Y float64
 }
